+++ examples/web-api/api.php
@@ -4,12 +4,56 @@ require './vendor/autoload.php';
 
 use function Concise\app;
 use function Concise\Routing\route;
+use function Concise\Routing\post;
 
 use function Concise\Http\Response\response;
+use function Concise\Http\Request\path as requestPath;
 use function Concise\Http\Response\setHeader;
+use function Concise\Http\Response\sta;
 use function Concise\Http\Response\send;
+use function Concise\Http\Response\statusCode;
 use function Concise\Http\Session\set as setSession;
 use function Concise\FP\ifElse;
+use function Concise\FP\tryCatch;
+
+use function Concise\Middleware\Factory\create as createMiddleware;
+
+function validTokens()
+{
+  return [
+    'Bearer abcd1234efgh5678',
+    'Bearer 1234abcd5678efgh'
+  ];
+}
+
+function authorizer($nextRouteHandler)
+{
+  return tryCatch(function ($request) use ($nextRouteHandler) {
+    if (!isset($_SERVER['HTTP_AUTHORIZATION']) || !in_array($_SERVER['HTTP_AUTHORIZATION'], validTokens())) {
+      throw new \Exception('Invalid access token');
+    }
+    return $nextRouteHandler($request);
+  }, function (\Exception $error) {
+    return send(setHeader('Content-Type', 'application/json')(response(json_encode([
+      'error' => true,
+      'status' => 401,
+      'message' => $error->getMessage()
+    ]))(statusCode(401, []))));
+  });
+}
+
+function routeProtected($requestPath)
+{
+  return substr($requestPath, 0, 4) === '/api' && strpos($requestPath, 'login') === false;
+}
+
+$authMiddleware = createMiddleware(function (callable $nextRouteHandler, array $middlewareParams = [], array $request = []) {
+  return ifElse('routeProtected', function () use ($nextRouteHandler, $request) {
+    return authorizer($nextRouteHandler)($request);
+  }, function () use ($nextRouteHandler, $request) {
+    return $nextRouteHandler($request);
+  })(requestPath());
+});
 
 app([
   route('GET', '/home', function () {
@@ -19,10 +63,10 @@ app([
     ]), [])));
   }),
 
-  route('GET', '/api/user/:id/order', function ($params) {
+  route('GET', '/api/user/:id', function (array $request) {
     return send(setHeader('Content-Type', 'application/json')(response(json_encode([
       'route' => '/api/user',
-      'data'  => [ 'user' => [ 'id' => $params['id'] ] ]
+      'data'  => [ 'user' => [ 'id' => $request['params']['id'] ] ]
     ]), [])));
   }),
 
@@ -33,27 +77,28 @@ app([
     ]), [])));
   }),
 
-  route('GET', '/api/upload/:upload_id', function ($params) {
+  route('GET', '/api/upload/:upload_id', function (array $request) {
     return send(setHeader('Content-Type', 'application/json')(response(json_encode([
       'route' => 'GET upload with ID',
-      'data'  => [ 'user' => [ 'upload_id' => $params['upload_id'] ] ]
+      'data'  => [ 'user' => [ 'upload_id' => $request['params']['upload_id'] ] ]
     ]), [])));
   }),
 
-  route('GET', '/api/auth/:password', function ($params) {
-    return send(ifElse(function ($password) {
-      return $password === 'V3ryS3cur3Passw0rd';
+  post('/api/auth/login')(function (array $request) {
+    return send(ifElse(function ($body) {
+      return isset($body['password']) && $body['password'] === 'V3ryS3cur3Passw0rd';
     })(function () {
       return setHeader('Content-Type', 'application/json')(response(json_encode([
-        'route' => 'GET upload with ID',
         'error' => false,
-        'accesstoken'  => setSession([ 'accesstoken' => '12v6gh5y643fds453ghgdf4zmb7439kl'])['accesstoken']
+        'access_token'  => setSession([ 'access_token' => '12v6gh5y643fds453ghgdf4zmb7439kl'])['access_token']
       ]), []));
     })(function () {
-      return setHeader('Content-Type', 'application/json')(response(json_encode([
+      return statusCode(422)(setHeader('Content-Type', 'application/json')(response(json_encode([
         'error'   => true,
-        'message' => 'Invalid auth code'
-      ]), []));
-    })($params['password']));
+        'message' => 'Invalid password'
+      ]), [])));
+    })($request['body']));
   })
-], []);
+], [
+  $authMiddleware
+]);