+++ spec/Concise/appTest.php
@@ -4,6 +4,9 @@ use PHPUnit\Framework\TestCase;
 use TestUtils\Spy;
 use function Concise\app as app;
 use function Concise\Middleware\Factory\create as createMiddleware;
+use function Concise\Http\Response\response;
+use function Concise\Http\Response\statusCode;
+use function Concise\Http\Response\send;
 
 function appMockRoutes()
 {
@@ -37,16 +40,30 @@ function appMockRoutes()
 
 class AppTest extends TestCase
 {
+  public function setup()
+  {
+    parent::setup();
+
+    $_POST = [];
+    $_GET = [];
+  }
+
   public function testAppRouteHandlerCalledWhenRouteExists()
   {
     $_SERVER['HTTP_HOST']= 'google.com';
     $_SERVER['REQUEST_URI']= '/api/user/20';
     $_SERVER['REQUEST_METHOD'] = 'DELETE';
 
-    $deleteHandlercallCount = 0;
-    $handlerCallArgs = [];
-    $deleteHandlerSpy = (new Spy())->setFunction(function () {
-    });
+    $expectedRequest = [
+      'params' => [
+        'id' => '20'
+      ],
+      'method' => 'DELETE',
+      'query' => [],
+      'body' => []
+    ];
+
+    $deleteHandlerSpy = Spy::create();
 
     $routes = appMockRoutes();
     $deleteUserRoute = [
@@ -64,9 +81,7 @@ class AppTest extends TestCase
     $handlerFirstCallFirstArg = $deleteHandlerSpy->getCall(0)[0];
 
     $this->assertEquals(1, $deleteHandlerSpy->callCount());
-    $this->assertEquals([
-      'id' => '20'
-    ], $handlerFirstCallFirstArg);
+    $this->assertEquals($expectedRequest, $handlerFirstCallFirstArg);
   }
 
   public function testAppRouteNotFound()
@@ -127,10 +142,17 @@ class AppTest extends TestCase
     $_SERVER['REQUEST_URI']= '/api/user/20/orders/100';
     $_SERVER['REQUEST_METHOD'] = 'DELETE';
 
-    $middlewareHandlercallCount = 0;
-    $handlerCallArgs = [];
-    $middlewareHandlerSpy = (new Spy())->setFunction(function () {
-    });
+    $expectedRequest = [
+      'params' => [
+        'id' => '20',
+        'order_id' => '100'
+      ],
+      'method' => 'DELETE',
+      'query' => [],
+      'body' => []
+    ];
+
+    $middlewareHandlerSpy = Spy::create();
 
     $mockMiddleware = createMiddleware($middlewareHandlerSpy['spy']);
 
@@ -151,10 +173,7 @@ class AppTest extends TestCase
     $middlewareFirstCallThirdArg = $middlewareHandlerSpy->getCall(0)[2];
 
     $this->assertEquals(1, $middlewareHandlerSpy->callCount());
-    $this->assertEquals([
-      'id' => '20',
-      'order_id' => '100'
-    ], $middlewareFirstCallThirdArg);
+    $this->assertEquals($expectedRequest, $middlewareFirstCallThirdArg);
   }
 
   public function testMultipleMiddlewaresInvokedWhenRouteMatches()
@@ -163,16 +182,26 @@ class AppTest extends TestCase
     $_SERVER['REQUEST_URI']= '/api/user/20/orders/100';
     $_SERVER['REQUEST_METHOD'] = 'DELETE';
 
-    $expectedParams = [
-      'id' => '20',
-      'order_id' => '100'
+    $expectedRequest = [
+      'params' => [
+        'id' => '20',
+        'order_id' => '100'
+      ],
+      'method' => 'DELETE',
+      'query' => [],
+      'body' => []
     ];
 
-    $firstMiddlewareHandler = (new Spy())->setFunction(function (callable $nextRouteHandler, array $middlewareParams = array(), array $reqParams = array()) {
-      return $nextRouteHandler($reqParams);
+    $firstMiddlewareHandler = Spy::create(function (callable $nextRouteHandler, array $middlewareParams = array(), array $request = array()) {
+      return $nextRouteHandler($request);
     });
 
-    $secondMiddlewareHandler = (new Spy())->setFunction(function () {
+    $secondMiddlewareHandler = Spy::create(function (callable $nextRouteHandler, array $middlewareParams = array(), array $request = array()) {
+      return $nextRouteHandler($request);
+    });
+
+    $routeHandler = Spy::create(function (array $request) {
+      return send(response('OK')(statusCode(201, [])));
     });
 
     $firstMockMiddleware = createMiddleware($firstMiddlewareHandler['spy']);
@@ -183,21 +212,29 @@ class AppTest extends TestCase
       [
         'method' => 'DELETE',
         'pattern' => '/api/user/:id/orders/:order_id',
-        'handler' => function () {
-        },
+        'handler' => $routeHandler['spy'],
         'regex'   => '/\/api\/user\/(?<id>[\w\-]+)\/orders\/(?<order_id>[\w\-]+)/',
-        'params'   => ['id']
+        'params'   => ['id', 'order_id']
       ]
     ];
 
-    app(array_merge($routes, $deleteUserRoute), [ $firstMockMiddleware, $secondMockMiddleware ]);
+    $response = app(array_merge($routes, $deleteUserRoute), [ $firstMockMiddleware, $secondMockMiddleware ]);
 
     $firstMiddlewareFirstCallThirdArg = $firstMiddlewareHandler->getCall(0)[2];
     $secondMiddlewareFirstCallThirdArg = $secondMiddlewareHandler->getCall(0)[2];
 
     $this->assertEquals(1, $firstMiddlewareHandler->callCount(), 'should call the first middleware once');
     $this->assertEquals(1, $secondMiddlewareHandler->callCount(), 'should call the second middleware once');
-    $this->assertEquals($expectedParams, $firstMiddlewareFirstCallThirdArg, 'should call the first middleware with correct params');
-    $this->assertEquals($expectedParams, $secondMiddlewareFirstCallThirdArg, 'should call the second middleware with correct params');
+    $this->assertEquals(1, $routeHandler->callCount(), 'should call the route handler');
+
+    $this->assertEquals($expectedRequest, $firstMiddlewareFirstCallThirdArg, 'should call the first middleware with correct params');
+    $this->assertEquals($expectedRequest, $secondMiddlewareFirstCallThirdArg, 'should call the second middleware with correct params');
+    $this->assertEquals([
+      'headers' => [
+        'Content-Type' => 'text/html'
+      ],
+      'body' => 'OK',
+      'status' => 201
+    ], $response);
   }
 }